requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
arabic-reshaper==3.0.0
argilla==0.0.1
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==24.2.0
azure-core==1.30.2
azure-storage-blob==12.19.0
backoff==2.2.1
base58==2.1.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.6.2
build==1.2.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.5
chromadb==0.5.4
click==8.1.6
coincurve==20.0.0
colorama==0.4.6
coloredlogs==15.0.1
Colorthon==2.9.6
contourpy==1.1.1
crispy-bootstrap4==2022.1
cryptography==42.0.7
cssselect2==0.7.0
cycler==0.12.1
Cython==3.0.9
cytoolz==0.12.3
dataclasses-json==0.6.7
datasets==2.14.2
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.7
distro==1.9.0
Django==4.2.5
django-bower==5.2.0
django-crispy-forms==2.0
dnspython==2.6.1
docopt==0.6.2
ecdsa==0.19.0
email_validator==2.2.0
et-xmlfile==1.1.0
eth-abi==2.2.0
eth-account==0.4.0
eth-hash==0.7.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
Faker==22.0.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.12.2
filetype==1.2.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flatbuffers==24.3.25
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.6.0
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.22.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-generativeai==0.8.3
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
hexbytes==0.3.1
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.4
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflect==7.2.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.1.2
langchain-community==0.0.18
langchain-core==0.3.13
langchain-google-genai==2.0.1
langchain-openai==0.0.5
langchain-text-splitters==0.2.2
langsmith==0.1.137
lapx==0.5.5
lxml==5.2.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.3
matplotlib==3.8.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysqlclient==2.2.0
network==0.1
networkx==3.1
nltk==3.8.1
num2words==0.5.13
numpy==1.25.2
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.18.1
openai==1.35.13
opencv-python==4.8.0.74
openpyxl==3.1.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.6
oscrypto==1.3.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.0.3
parsimonious==0.8.1
pdf2image==1.17.0
pdfkit==1.0.0
pdfminer.six==20240706
Pillow==10.0.1
pip==24.2
platformdirs==3.11.0
posthog==3.5.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
pymongo==4.6.1
PyMuPDF==1.24.12
pynvml==11.5.3
pypandoc==1.13
pyparsing==3.1.1
pypdf==4.2.0
PyPDF2==3.0.1
PyPika==0.48.9
pypng==0.20220715.0
pyproject_hooks==1.1.0
pyreadline3==3.4.1
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
regex==2023.6.3
render==1.0.0
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.1
rlp==1.2.0
rsa==4.9
safetensors==0.4.5
schedule==1.2.1
selenium==4.25.0
sentencepiece==0.1.99
setuptools==65.5.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.4.4
starlette==0.37.2
svglib==1.5.1
sympy==1.12
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
text2sql==0.1.1
tiktoken==0.5.2
tinycss2==1.3.0
tokenizers==0.19.1
toolz==0.12.1
torch==2.1.0
tqdm==4.65.0
transformers==4.44.2
trio==0.27.0
trio-websocket==0.11.1
tronapi==3.1.6
tronpy==0.5.0
trx-utils==1.0.4
typeguard==4.2.1
typer==0.12.3
typing_extensions==4.11.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.2
ujson==5.10.0
unstructured==0.7.5
uritemplate==4.1.1
uritools==4.0.2
urllib3==1.26.16
uvicorn==0.30.1
virtualenv==20.24.5
watchfiles==0.22.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wheel==0.41.0
wrapt==1.16.0
wsproto==1.2.0
xhtml2pdf==0.2.15
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.3.0
yarl==1.9.2
zipp==3.19.2
